+++ packages/react/src/hooks/useOrganizations.ts
@@ -1,14 +1,22 @@
-import { CreateOrganizationParams, OrganizationResource } from '@clerk/types';
+import {
+  CreateOrganizationParams,
+  OrganizationMembershipResource,
+  OrganizationResource,
+} from '@clerk/types';
 import { useContext } from 'react';
 
 import { assertWrappedByClerkProvider } from '../contexts/assertHelpers';
 import { StructureContext } from '../contexts/StructureContext';
-import { useClerk } from './useClerk';
+import { useClerk } from '../hooks';
 
 type UseOrganizations = {
-  createOrganization: (params: CreateOrganizationParams) => Promise<OrganizationResource>;
-  getOrganizations: () => Promise<OrganizationResource[]>;
-  getOrganization: (organizationId: string) => Promise<OrganizationResource | undefined>;
+  createOrganization: (
+    params: CreateOrganizationParams,
+  ) => Promise<OrganizationResource>;
+  getOrganizationMemberships: () => Promise<OrganizationMembershipResource[]>;
+  getOrganization: (
+    organizationId: string,
+  ) => Promise<OrganizationResource | undefined>;
 };
 
 export function useOrganizations(): UseOrganizations {
@@ -18,7 +26,7 @@ export function useOrganizations(): UseOrganizations {
 
   return {
     createOrganization: clerk.createOrganization,
-    getOrganizations: clerk.getOrganizations,
+    getOrganizationMemberships: clerk.getOrganizationMemberships,
     getOrganization: clerk.getOrganization,
   };
 }