+++ packages/nextjs/CHANGELOG.md
@@ -42,6 +42,22 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+### [2.11.13](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@2.11.12...@clerk/nextjs@2.11.13) (2022-03-09)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+
+
+
+
+### [2.11.12](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@2.11.12-staging.0...@clerk/nextjs@2.11.12) (2022-03-09)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+
+
+
+
 ### [2.11.10](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@2.11.9...@clerk/nextjs@2.11.10) (2022-03-04)
 
 **Note:** Version bump only for package @clerk/nextjs