+++ packages/shared/CHANGELOG.md
@@ -23,6 +23,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+### [0.0.12](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.0.12-staging.0...@clerk/shared@0.0.12) (2022-03-09)
+
+**Note:** Version bump only for package @clerk/shared
+
+
+
+
+
 ### [0.0.10](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.0.9...@clerk/shared@0.0.10) (2022-03-04)
 
 **Note:** Version bump only for package @clerk/shared