+++ packages/types/CHANGELOG.md
@@ -32,6 +32,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+### [1.28.3](https://github.com/clerkinc/javascript/compare/@clerk/types@1.28.3-staging.0...@clerk/types@1.28.3) (2022-03-09)
+
+**Note:** Version bump only for package @clerk/types
+
+
+
+
+
 ### [1.28.1](https://github.com/clerkinc/javascript/compare/@clerk/types@1.28.0...@clerk/types@1.28.1) (2022-03-04)